Update README.md
What issue did they want you to address? Artemis Financial is a company that is invested in their clients financial future needs and security. Their software requirement were to perform a vulnerability report for their current software and list the areas of security needing work.
I performed interpreting the clients needs well while identifying the needs of Artemis Financial. It is important to code securely because it helps prevent breaches in software and leads to a more respected company when clients are looking to secure their assets.
What I found helpful while working through the vulnerability assessment was running through the dependency check and looking through the list of vulnerabilities.
I used the list from the dependency check to increase layers of security as well as a manual review of the code which is useful for future projects as well.
How did you ensure the code and software application were functional and secure? After refactoring code, how did you check to see whether you introduced new vulnerabilities? To ensure the code and application were functional and secure I performed the dependency checks and also a checksum. After adding a suppression.xml file to suppress false positive, I reran the dependency check to ensure I did not introduce any new vulnerabilities.
When executing future tasks or projects I will use the dependency check and checksum tools as well as the self-signed certificate.
I would like future employers to see that I can interpret the clients needs for if you can't meet those then the client won't be happy with the product. I would also like future employers to see that I can perform a vulnerability assessment and give a list of vulnerabilities that could cause the clients issues. To be able to notice false positives and suppress them for clients as well.